Move Dockerfile and Makefile to flask directory
