+++ pollex-flask/Makefile
@@ -46,7 +46,7 @@ MAKE = make
 
 # Sources
 
-sources = index.py templates/*.html static/*.css
+sources = index.py # templates/*.html static/*.css
 
 # Do it all
